services/monitoring-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0

